+++ wolfssl/wolfcrypt/types.h
@@ -441,7 +441,7 @@
     #ifdef WOLFSSL_RIOT_OS
         #define EXIT_TEST(ret) exit(ret)
     #elif defined(HAVE_STACK_SIZE)
-        #define EXIT_TEST(ret) return (void*)((long)(ret))
+        #define EXIT_TEST(ret) return (void*)((size_t)(ret))
     #else
         #define EXIT_TEST(ret) return ret
     #endif
